+++ src/components/Blog/BlogList/BlogListItem/BlogListItem.test.tsx
@@ -14,6 +14,17 @@ const testProps: BlogListItemProps = {
   date: new Date(2022, 7, 22).toISOString(),
 };
 
+const testPropsWebinar: BlogListItemProps = {
+  title: "Item title",
+  titleTag: "h3",
+  summary: "Item snippet",
+  slug: "item-slug",
+  contentType: "webinar",
+  category: { title: "Curriculum Planning", slug: "curriculum-planning" },
+  mainImage: "stringvideoplaybackid",
+  date: new Date(2022, 7, 22).toISOString(),
+};
+
 describe("components/BlogListItem", () => {
   test("renders the correct heading tag", () => {
     const { getByRole } = renderWithTheme(
@@ -44,7 +55,7 @@ describe("components/BlogListItem", () => {
   });
   test("webinar: button should have the correct href", async () => {
     const { getByRole } = renderWithTheme(
-      <BlogListItem {...testProps} contentType="webinar" />
+      <BlogListItem {...testPropsWebinar} />
     );
     const button = getByRole("link", { name: testProps.title });
     expect(button).toHaveAttribute("href", `/beta/webinars/${testProps.slug}`);
@@ -52,7 +63,7 @@ describe("components/BlogListItem", () => {
 
   test("webinar: should contain link to category", async () => {
     const { getByRole } = renderWithTheme(
-      <BlogListItem {...testProps} contentType="webinar" />
+      <BlogListItem {...testPropsWebinar} />
     );
     const button = getByRole("link", { name: testProps.category.title });
     expect(button).toHaveAttribute(