+++ src/components/Blog/BlogWebinarsListAndCategories.tsx
@@ -49,11 +49,12 @@ const BlogWebinarsListAndCategories: FC<BlogWebinarsListAndCategoriesProps> = (
       </GridArea>
       {/* @todo is there a nicer way to make this 1 column spacer? */}
       <GridArea $order={1} $colSpan={[12, 1]} />
-      <GridArea $order={[1, 0]} $colSpan={[12, 7, 8]} $mt={[48, 72]}>
+      <GridArea $order={[1, 0]} $colSpan={[12, 7, 8]} $mt={[16, 72]}>
         <BlogList
           items={blogs}
           withContainingHrs
           withPagination
+          withImage={page === "blog-index" ? false : true}
           withUpcomingItem
         />
       </GridArea>