+++ src/pages/beta/webinars/[webinarSlug].tsx
@@ -4,44 +4,59 @@ import {
   GetStaticPropsResult,
   NextPage,
 } from "next";
-import { PortableText } from "@portabletext/react";
+import { uniqBy } from "lodash/fp";
 
 import { getSeoProps } from "../../../browser-lib/seo/getSeoProps";
 import Layout from "../../../components/Layout";
-import { Heading } from "../../../components/Typography";
 import CMSClient from "../../../node-lib/cms";
-import { Webinar } from "../../../common-lib/cms-types";
+import { TeamMemberPreview, Webinar } from "../../../common-lib/cms-types";
+import { getBlogWebinarPostBreadcrumbs } from "../../../components/Breadcrumbs/getBreadcrumbs";
+import Box from "../../../components/Box";
 import { decorateWithIsr } from "../../../node-lib/isr";
+import BlogPortableText from "../../../components/Blog/BlogPortableText/BlogPortableText";
+import Flex from "../../../components/Flex";
+import BlogWebinarsIndexLayout from "../../../components/Blog/BlogWebinarsIndexLayout";
+import WebinarVideo from "../../../components/Blog/WebinarVideo";
+import { BlogJsonLd } from "../../../browser-lib/seo/getJsonLd";
 
 export type SerializedWebinar = Omit<Webinar, "date"> & {
   date: string;
+  author: TeamMemberPreview | undefined;
 };
 
 export type WebinarPageProps = {
   webinar: SerializedWebinar;
+  categories: { title: string; slug: string }[];
 };
 
-/**
- * @TODO: Remove /webinars/* from next-sitemap.config.js when built
- */
-
 const WebinarDetailPage: NextPage<WebinarPageProps> = (props) => {
+  const { webinar, categories } = props;
+
   return (
-    <Layout seoProps={getSeoProps(props.webinar.seo)} $background="grey1">
-      <Heading tag="h1" $font="heading-5">
-        {props.webinar.title}
-      </Heading>
-      {props.webinar.date} <br />
-      Hosted by: {props.webinar.hosts?.map((host) => host?.name).join(", ")}
-      <p>
-        An example of rich text via the <code>summaryPortableText</code> field
-      </p>
-      <div style={{ border: "1px solid red" }}>
-        <PortableText
-          value={props.webinar.summaryPortableText}
-          components={{}}
-        />
-      </div>
+    <Layout
+      seoProps={getSeoProps({
+        ...props.webinar.seo,
+        title: webinar.seo?.title || webinar.title,
+        description: webinar.seo?.description,
+        imageUrl: "", // @TODO: add image from video frame
+      })}
+      $background="white"
+      breadcrumbs={getBlogWebinarPostBreadcrumbs(
+        categories,
+        webinar,
+        "beta/webinars",
+        "Webinars"
+      )}
+    >
+      <BlogWebinarsIndexLayout content={props}>
+        <Flex $position={"relative"} $mt={56}>
+          <WebinarVideo webinar={webinar} />
+        </Flex>
+        <Box $mt={[48]}>
+          <BlogPortableText portableText={webinar.summaryPortableText} />
+        </Box>
+      </BlogWebinarsIndexLayout>
+      <BlogJsonLd blog={props.webinar} />
     </Layout>
   );
 };
@@ -78,14 +93,23 @@ export const getStaticProps: GetStaticProps<
     };
   }
 
-  const webinar = {
+  const webinarResults = await CMSClient.webinars();
+
+  const categories = uniqBy(
+    "title",
+    webinarResults.map((w) => w.category)
+  ).sort((a, b) => (a.title < b.title ? -1 : 1));
+
+  const webinar: SerializedWebinar = {
     ...webinarResult,
     date: webinarResult.date.toISOString(),
+    author: webinarResult.hosts[0], // make the first host equivalent to a blog author
   };
 
   const results: GetStaticPropsResult<WebinarPageProps> = {
     props: {
       webinar,
+      categories,
     },
   };
   const resultsWithIsr = decorateWithIsr(results);