+++ src/pages/index.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import { toPlainText } from "@portabletext/react";
 
 import CMSClient from "../node-lib/cms";
-import { HomePage, WebinarPreview } from "../common-lib/cms-types";
+import { HomePage } from "../common-lib/cms-types";
 import { decorateWithIsr } from "../node-lib/isr";
 import { getSeoProps } from "../browser-lib/seo/getSeoProps";
 import Grid from "../components/Grid";
@@ -384,26 +384,22 @@ export const getStaticProps: GetStaticProps<HomePageProps> = async (
     limit: 5,
   });
 
-  // const webinarResults = await CMSClient.webinars({
-  //   previewMode: isPreviewMode,
-  //   limit: 5,
-  // });
-  const webinarResults: WebinarPreview[] = [];
-
   const blogPosts = blogResults.map((blog) => ({
     ...blog,
     type: "blog-post" as const,
   }));
 
-  const webinars = webinarResults.map((webinar) => ({
-    ...webinar,
-    type: "webinar" as const,
-  }));
+  // @todo add to posts array and un-comment when webinars are finshed
+  // const webinarResults = await CMSClient.webinars({
+  //   previewMode: isPreviewMode,
+  //   limit: 5,
+  // });
+  // const webinars = webinarResults.map((webinar) => ({
+  //   ...webinar,
+  //   type: "webinar" as const,
+  // }));
 
-  const posts = [...blogPosts, ...webinars]
-    .sort(sortByDate)
-    .slice(0, 4)
-    .map(serializeDate);
+  const posts = [...blogPosts].sort(sortByDate).slice(0, 4).map(serializeDate);
 
   const results: GetStaticPropsResult<HomePageProps> = {
     props: {