+++ src/styles/theme/types.ts
@@ -56,6 +56,7 @@ export type PixelSpacing =
   | 300
   | 320
   | 360
+  | 400
   | 480
   | 600
   | 720